+++ CHANGELOG.md
@@ -2,6 +2,7 @@ See also http://pvp.haskell.org/faq
 
 ### 0.1.4.0
 
+* New module `Network.Http.Client.WebSocket` providing basic RFC6455 support.
 * New function 'inputStreamBodyChunked' supporting breaking up over-sized chunks.
 
 ### 0.1.3.0